Разбил функции на модули
